requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
sqlalchemy[asyncio]>=2.0.30
asyncpg>=0.29.0
psycopg2-binary>=2.9.9
alembic>=1.13.1
pydantic[email]>=2.7.1
pydantic-settings>=2.2.1
celery>=5.4.0
redis>=5.0.4
boto3>=1.34.0
python-docx>=1.1.2
python-multipart>=0.0.9
anthropic>=0.26.1
httpx>=0.27.0
stripe>=9.9.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
structlog>=24.1.0
sentry-sdk[fastapi]>=2.0.0
aioboto3>=12.0.0
pdfminer.six>=20221105
pymupdf>=1.24.0
cryptography>=42.0.0
twilio>=9.0.0
